+++ src/test/kotlin/blackjack/BlackjackShoeTest.kt
@@ -1,5 +1,6 @@
 package blackjack
 
+import blackjack.domain.BlackjackShoe
 import io.kotest.assertions.throwables.shouldThrow
 import org.junit.jupiter.api.Test
