+++ src/test/kotlin/blackjack/RankTest.kt
@@ -1,5 +1,6 @@
 package blackjack
 
+import blackjack.domain.Rank
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
