+++ src/test/kotlin/blackjack/SuitTest.kt
@@ -1,11 +1,10 @@
 package blackjack
 
+import blackjack.domain.Suit
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.inspectors.shouldForAll
 import io.kotest.matchers.shouldBe
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.ValueSource
 
 class SuitTest {
     @Test
